modules/mbapi/requirements.txt
Flask==3.0.0
kafka-python==2.0.6
Werkzeug==3.1.3
flask_restx==1.3.0
SQLAlchemy==2.0.39